Move bind into Tcp type
